apps: move global libctx and property query into their own file
The header has been split out so the functions should be as well.
