+++ apps/include/app_libctx.h
@@ -10,6 +10,8 @@
 #ifndef OSSL_APPS_LIBCTX_H
 # define OSSL_APPS_LIBCTX_H
 
+#include <openssl/types.h>
+
 OSSL_LIB_CTX *app_create_libctx(void);
 OSSL_LIB_CTX *app_get0_libctx(void);
 int app_set_propq(const char *arg);
