+++ apps/lib/build.info
@@ -10,7 +10,7 @@ ENDIF
 # Source for libapps
 $LIBAPPSSRC=apps.c apps_ui.c opt.c fmt.c s_cb.c s_socket.c app_rand.c \
         columns.c app_params.c names.c app_provider.c app_x509.c http_server.c \
-        engine.c engine_loader.c
+        engine.c engine_loader.c app_libctx.c
 
 IF[{- !$disabled{apps} -}]
   LIBS{noinst}=../libapps.a
